fix(core): move detectInvokedPackageManager off of tao
