+++ packages/tao/src/shared/package-manager.ts
@@ -86,32 +86,3 @@ export function getPackageManagerVersion(
 ): string {
   return execSync(`${packageManager} --version`).toString('utf-8').trim();
 }
-
-/**
- * Detects which package manager was used to invoke create-nx-{plugin|workspace} command
- * based on the main Module process that invokes the command
- * - npx returns 'npm'
- * - pnpx returns 'pnpm'
- * - yarn create returns 'yarn'
- *
- * Default to 'npm'
- */
-export function detectInvokedPackageManager(): PackageManager {
-  let detectedPackageManager: PackageManager = 'npm';
-  // mainModule is deprecated since Node 14, fallback for older versions
-  const invoker = require.main || process['mainModule'];
-
-  // default to `npm`
-  if (!invoker) {
-    return detectedPackageManager;
-  }
-
-  for (const pkgManager of ['pnpm', 'yarn', 'npm'] as const) {
-    if (invoker.path.includes(pkgManager)) {
-      detectedPackageManager = pkgManager;
-      break;
-    }
-  }
-
-  return detectedPackageManager;
-}
